Slack Server/requirements.txt
Flask==2.2.2
Flask_Cors==4.0.0
python-dotenv==1.0.0
Requests==2.31.0
slack_sdk==3.21.3